+++ demo/button.gd
@@ -1,8 +1,8 @@
 extends Button
 
 export(String) var scene
-export(float) var fade_out_speed = 1
-export(float) var fade_in_speed = 1
+export(float) var fade_out_speed = 1.0
+export(float) var fade_in_speed = 1.0
 export(String) var fade_out_pattern = "fade"
 export(String) var fade_in_pattern = "fade"
 export(float, 0, 1) var fade_out_smoothness = 0.1
@@ -10,7 +10,7 @@ export(float, 0, 1) var fade_in_smoothness = 0.1
 export(bool) var fade_out_inverted = false
 export(bool) var fade_in_inverted = false
 export(Color) var color = Color(0, 0, 0)
-export(float) var timeout = 0
+export(float) var timeout = 0.0
 export(bool) var clickable = false
 
 onready var fade_out_options = SceneManager.create_options(fade_out_speed, fade_out_pattern, fade_out_smoothness, fade_out_inverted)